Added tabs to page sale
