+++ src/langs/sq.json
@@ -231,5 +231,8 @@
 	"almostOutOfSpaceText": "Nëse ju mbaron hapësira ruajtëse, nuk do të mund të merrni njoftime!",
 	"outOfSpaceText": "Nuk do të merrni më njoftime, ju lutemi kontaktoni pronarin për të rritur numrin e njoftimeve!",
 	"used": "të përdorura",
-	"of": "nga"
+	"of": "nga",
+	"tab": "Faqja",
+	"addTab": "Shto faqe",
+	"maxTabs": "Numri maksimal i faqeve është 5"
 }
